Getting a table of general information for a minileague's current gw

This function returns a table of general information for a minileague. Automatically return a table for the current gameweek.

get_league_general_info(league_number)

Arguments

league_number

the league number you want a table for

Examples

# Getting the minileague information
get_league_general_info(570437)
#>            entry_name points total_points overall_rank rank_change
#> 1     Erik Ten Haggis     54          823       296727      136771
#> 2     Cookie Monstars     52          821       324665      199677
#> 3           Yes Ndidi     61          804       623936      104977
#> 4         Martial Law     56          788      1034306      261936
#> 5     Rise of Haaland     70          774      1477517     -185706
#> 6 Route-One Direction     64          733      3114166      -50761
#> 7     Humza's hunneez     37          709      4111911     1096328
#> 8            SROSS FC     43          709      4112508      823913
#> 9         Newhaven FC     36          632      6801888      627796